src/data/sec_filings.py
import os
from typing import Dict, List, Optional
import requests
from bs4 import BeautifulSoup
from datetime import datetime
import pandas as pd
from dotenv import load_dotenv

class SECFilingsFetcher:
    """Class to fetch and parse SEC filings (10-K, 10-Q)"""
    
    def __init__(self):
        load_dotenv()
        self.api_key = os.getenv('SEC_API_KEY')
        self.base_url = "https://api.sec-api.io"
        
    def get_company_filings(self, ticker: str, form_type: str = "10-K") -> List[Dict]:
        """
        Fetch SEC filings for a given company ticker
        
        Args:
            ticker (str): Company ticker symbol
            form_type (str): Type of filing (10-K, 10-Q, etc.)
            
        Returns:
            List[Dict]: List of filing information
        """
        if not self.api_key:
            raise ValueError("SEC API key not found. Please set SEC_API_KEY in .env file")
            
        headers = {
            "Authorization": f"Bearer {self.api_key}",
            "Content-Type": "application/json"
        }
        
        # TODO: Implement actual API call
        # This is a placeholder for the actual implementation
        return []
    
    def parse_filing(self, filing_url: str) -> Dict:
        """
        Parse the content of an SEC filing
        
        Args:
            filing_url (str): URL to the SEC filing
            
        Returns:
            Dict: Parsed filing content
        """
        # TODO: Implement filing parsing logic
        return {}
    
    def extract_financial_metrics(self, filing_content: Dict) -> pd.DataFrame:
        """
        Extract key financial metrics from filing content
        
        Args:
            filing_content (Dict): Parsed filing content
            
        Returns:
            pd.DataFrame: DataFrame containing financial metrics
        """
        # TODO: Implement financial metrics extraction
        return pd.DataFrame()

if __name__ == "__main__":
    # Example usage
    fetcher = SECFilingsFetcher()
    # Add example code here 