requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.36
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
ta-lib>=0.4.28
requests>=2.31.0
beautifulsoup4>=4.12.0
sec-api>=1.0.0
python-dotenv>=1.0.0
mplfinance>=0.12.10b0 